+++ ewoh-spark-app/server/app.module.ts
@@ -36,6 +36,7 @@ import { AasModule } from './modules/aas/aas.module';
 import { TracingModule } from './modules/tracing/tracing.module';
 import { TracingInterceptor } from './modules/tracing/tracing.interceptor';
 import { WorkOrchestrationModule } from './modules/work-orchestration/work-orchestration.module';
+import { ObservabilityModule } from './modules/observability/observability.module';
 
 @Module({
   imports: [
@@ -71,6 +72,7 @@ import { WorkOrchestrationModule } from './modules/work-orchestration/work-orche
     AasModule,
     TracingModule,
     WorkOrchestrationModule,
+    ObservabilityModule,
     // ====== @route-section: business-modules END ======
 
     // ⚠️ @route-order: last